Make sure load/half is integer

+++ sortedcontainers/sortedlist.py
@@ -18,7 +18,7 @@ else:
 class SortedList(MutableSequence):
     def __init__(self, iterable=None, load=100):
         self.clear()
-        self._load, self._twice, self._half = load, load * 2, load / 2
+        self._load, self._twice, self._half = load, load * 2, int(load / 2)
 
         if iterable is not None:
             self.update(iterable)
@@ -678,7 +678,7 @@ class SortedList(MutableSequence):
             # Check load parameters.
 
             assert self._load >= 4
-            assert self._half == (self._load / 2)
+            assert self._half == int(self._load / 2)
             assert self._twice == (self._load * 2)
 
             # Check empty sorted list case.
